calorie-tracker/backend/requirements_advanced.txt
# Advanced Requirements for Resume-Worthy Calorie Tracker
# Core FastAPI and Database
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
cryptography==41.0.7

# Computer Vision and Machine Learning
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1

# Data Analysis and Visualization
pandas==2.1.3
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0
scipy==1.11.4

# API Integration and Web Scraping
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
selenium==4.15.2

# Caching and Performance
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0

# Two-Factor Authentication
pyotp==2.9.0
qrcode[pil]==7.4.2

# Testing and Quality Assurance
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# Code Quality and Documentation
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Monitoring and Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0
click==8.1.7

# File Processing and Storage
boto3==1.34.0
minio==7.2.0
python-magic==0.4.27

# Rate Limiting and Security
slowapi==0.1.9
backoff==2.2.1
python-json-logger==2.0.7

# Development Tools
ipython==8.17.2
jupyter==1.0.0
notebook==7.0.6

# Production Dependencies
gunicorn==21.2.0
supervisor==4.2.5
nginx==1.0.0

